+++ packages/mobx-binder-core/src/index.ts
@@ -6,8 +6,13 @@ export { AbstractField } from './model/fields/AbstractField'
 export { TextField } from './model/fields/TextField'
 export { ToggleField } from './model/fields/ToggleField'
 
+export { SimpleBinder, ErrorMessage } from './model/binder/SimpleBinder'
+
 export { Converter, AsyncConverter } from './conversion/Converter'
 export { TrimConverter } from './conversion/TrimConverter'
+
+export * from './validation/Validity'
 export * from './validation/Validator'
+
 export { ValidationError } from './conversion/ValidationError'
 export { isPromise } from './utils/isPromise'